requirements.txt
# Core dependencies
torch>=2.0.0
sentence-transformers>=2.2.2
qdrant-client>=1.7.0

# LLM Integration
anthropic>=0.39.0

# Code parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
